fix(stream-deck): 実装効率化
